+++ src/core_examples/config/layouts/local_vectorstore_adaptive_rag_config_graph.py
@@ -90,17 +90,17 @@ class LocalVectorStoreAdaptiveRAGConfigGraph(GraphLayout):
         self.GENERATION_NODE = SimpleNode(
             enhancer=GenerateAnswerAsyncInvoke(self.GENERARION_CHAIN),
             name=self.CONFIG_NODES["GENERATION_NODE"]["name"],
-            tags=[self.CONFIG_NODES["GENERATION_NODE"]["description"]],
+            metadata=self.CONFIG_NODES["GENERATION_NODE"]["metadata"],
         )
         self.RETRIEVER_NODE = SimpleNode(
             enhancer=RetrieveContextAsyncInvoke(self.RETRIEVER_CHAIN),
             name=self.CONFIG_NODES["RETRIEVER_NODE"]["name"],
-            tags=[self.CONFIG_NODES["RETRIEVER_NODE"]["description"]],
+            metadata=self.CONFIG_NODES["RETRIEVER_NODE"]["metadata"],
         )
         self.REWRITE_NODE = SimpleNode(
             enhancer=RewriteQuestionAsyncInvoke(self.REWRITE_CHAIN),
             name=self.CONFIG_NODES["REWRITE_NODE"]["name"],
-            tags=[self.CONFIG_NODES["REWRITE_NODE"]["description"]],
+            metadata=self.CONFIG_NODES["REWRITE_NODE"]["metadata"],
         )
 
         ## EDGES